Projeto final simples

+++ src/js/delete.js
@@ -1,13 +1,15 @@
-import { deleteUser} from "./api.js";
+import { deleteUser } from "./api.js";
 //deletar usuário
 
 export async function handleDelete(event){
   const button = event.target.closest(".delete-btn");
+    if(!button) return;
   const id = button.dataset.id;
-  
+
 
   console.log("id capturado:", id);
-  
+  const confirmou = confirm("Deletar usuário?");
+  if(!confirmou) return;
   try {
     await deleteUser(id);
